move resource parameters from spec to job
