+++ src/_canary/job.py
@@ -283,7 +283,7 @@ class Job(BaseJob):
         super().__init__()
         self.spec = spec
         self.workspace = workspace
-        self.rparameters = self.get_resource_parameters_from_spec()
+        self.rparameters = self.spec.rparameters
         pm = config.pluginmanager.hook
         self.launcher: Launcher = pm.canary_runtest_launcher(case=self)
         self._mask: Mask | None = None
@@ -823,40 +823,6 @@ class Job(BaseJob):
         variables["PATH"] = t.substitute(os.environ, missing="")
         return variables
 
-    def get_resource_parameters_from_spec(self) -> dict[str, int]:
-        resource_types: set[str] = set(config.resource_manager.types())
-        p = self.spec.parameters | self.spec.meta_parameters
-        rparameters: dict[str, int] = {}
-
-        for key in p.keys() & (resource_types | {"nodes"}):
-            value = p[key]
-            if not isinstance(value, int):
-                raise InvalidTypeError(key, value)
-            rparameters[key] = value
-        nodes = rparameters.get("nodes")
-        if nodes is None:
-            nodes = 1
-            for rtype, count in rparameters.items():
-                if rtype == "nodes":
-                    continue
-                if count <= 0:
-                    continue
-                slots_per_node = config.resource_manager.slots_per_node(rtype)
-                # CPU fallback for normal local cases.
-                if slots_per_node <= 0 and rtype in ("cpu", "cpus"):
-                    slots_per_node = cpu_count()
-                # If the resource is unknown/unavailable, leave node count alone.
-                # ResourceCapacityRule / ResourcePool.accommodates() will report
-                # the actual insufficiency later.
-                if slots_per_node <= 0:
-                    continue
-                nodes = max(nodes, ceil_div(count, slots_per_node))
-        rparameters["nodes"] = nodes
-        # Preserve default CPU/GPU resource parameters.
-        rparameters.setdefault("cpus", 1)
-        rparameters.setdefault("gpus", 0)
-        return rparameters
-
     def teardown(self) -> None:
         pass
 
@@ -978,11 +944,6 @@ def find_cache_dir(start: Path) -> Path | None:
     return None
 
 
-def ceil_div(a: int, b: int) -> int:
-    assert b != 0, "denominator must not be 0"
-    return (a + b - 1) // b
-
-
 def split_count(total: int, parts: int) -> list[int]:
     assert parts > 0
     q, r = divmod(total, parts)
@@ -991,9 +952,3 @@ def split_count(total: int, parts: int) -> list[int]:
 
 class MissingSourceError(Exception):
     pass
-
-
-class InvalidTypeError(Exception):
-    def __init__(self, name, value):
-        class_name = value.__class__.__name__
-        super().__init__(f"expected type({name})=type({value!r})=int, not {class_name}")
